scoutfs: export server commit holds
The calls for holding and applying commits in the server are currently
private.  The lock server is a server component that has been seperated
out into its own file.  Most of the time the server calls it during
commits so the btree changes made in the lock server are protected by
the commits.  But there are btree calls in the lock server that happen
outside of calls from the server.

Exporting these calls will let the lock server make all its btree
changes in server commits.

Signed-off-by: Zach Brown <zab@versity.com>